+++ CHANGES.rst
@@ -15,6 +15,8 @@ New Features
 
 - Catalog public API exposed [#3761, #3777, #3778, #3799, #3814, #3835, #3854, #3856, #3863, #3867, #3930, #3906, #3912, #3899, #3907, #3990]
 
+- Support in plot options to set which columns are visible for table viewers. [#4033]
+
 - Added `skewer` mode to footprint selection that only selects when clicking inside a footprint. [#3962]
 
 - Added ability to load catalogs from FITS file extensions through the Catalog loader. [#3998]